requirements.txt
Django==3.2.13
psycopg2-binary==2.9.3
pytest==7.0.0
python-dateutil==2.8.2